+++ src/Stash/Driver/Composite.php
@@ -40,7 +40,7 @@ class Composite extends AbstractDriver
      *
      * @throws \Stash\Exception\RuntimeException
      */
-    public function setOptions(array $options = array())
+    protected function setOptions(array $options = array())
     {
         $options += $this->getDefaultOptions();