+++ src/Stash/Driver/Redis.php
@@ -61,7 +61,7 @@ class Redis extends AbstractDriver
      *
      * @param array $options
      */
-    public function setOptions(array $options = array())
+    protected function setOptions(array $options = array())
     {
         $options += $this->getDefaultOptions();