+++ src/Stash/Driver/Sqlite.php
@@ -55,7 +55,7 @@ class Sqlite extends AbstractDriver
      *
      * @throws \Stash\Exception\RuntimeException
      */
-    public function setOptions(array $options = array())
+    protected function setOptions(array $options = array())
     {
         $options += $this->getDefaultOptions();