+++ tests/Stash/Test/Driver/ApcTest.php
@@ -26,8 +26,7 @@ class ApcTest extends AbstractDriverTest
         $options = $this->getOptions();
         $options['namespace'] = 'namespace_test';
         $options['ttl'] = 15;
-        $driver = new $driverType();
-        $driver->setOptions($options);
+        $driver = new $driverType($options);
 
         $this->assertAttributeEquals('namespace_test', 'apcNamespace', $driver, 'APC is setting supplied namespace.');
         $this->assertAttributeEquals(15, 'ttl', $driver, 'APC is setting supplied ttl.');